Address follow-up review comments
- TxCodeInputPopup: Use PopupLayout and Button components
- Logging: Use jsonToLog() instead of JSON.stringify()
- ProtocolSteps.ts: Extract runtime code to protocolStepInfo.ts
- Remove obsolete Dockerfile.e2e (config injection now via secrets)
